% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpling.R
\docType{package}
\name{corpling}
\alias{corpling}
\alias{corpling-package}
\title{corpling: Simple tools for corpus linguistics.}
\description{
Produces concordances of regular expressions and creates frequency
lists of ngrams.
}
\section{corpling functions}{

\code{\link{concord}},
\code{\link{load_txt}},
\code{\link{ngram_freq}}
}
\seealso{
\code{\link{load_txt}} loads files located in a directory on the hard
   drive into a list, each file in one element of the list.
   \code{\link{concord}} produces a concordance of a given regular
   expression and \code{\link{ngram_freq}} produces a frequency list of ngrams.
}

