% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_freq.R
\name{lead}
\alias{lead}
\title{Lead method used in ngram_freq.default}
\usage{
lead(x, n)
}
\arguments{
\item{x}{input character vector}

\item{n}{the number of words to lead by}
}
\description{
Lead method used in ngram_freq.default
}
\keyword{internal}

