% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_freq.R
\name{move_ahead}
\alias{move_ahead}
\title{Method used in ngram_freq.default}
\usage{
move_ahead(x, n)
}
\arguments{
\item{x}{input character vector}

\item{n}{the number of words to lead by}
}
\description{
Method used in ngram_freq.default
}
\keyword{internal}

