% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_freq.R
\name{ngram_freq}
\alias{ngram_freq}
\title{Produce a frequency list of ngrams.}
\usage{
ngram_freq(text, num_wd = 1, ignore_case = TRUE, order_by = "alpha",
  descending = FALSE, min_freq = 1, word_char = NULL)
}
\arguments{
\item{text}{The text with the ngrams whose frequencies are to be determined,
as either a character vector or something coercible to it, such as a list
of character vectors.}

\item{num_wd}{Specifies the number of words in the ngrams, whether single words
(\code{num_wd = 1}, the default), bigrams (\code{num_wd = 2}), trigrams
(\code{num_wd = 3}), or even larger ngrams.}

\item{ignore_case}{Specifies whether the frequency list be
case-insensitive (\code{ignore_case = TRUE}, the default) or case-sensitve
(\code{ignore_case = FALSE}). If case-insensitive, the ngrams are converted
to upper-case.}

\item{order_by}{Specifies whether the frequency list be ordered alphabetically
(\code{order_by = "alpha"}) or by frequency (\code{order_by = "freq"}).}

\item{descending}{Specifies whether the frequency list be ordered in
 ascending order (\code{descending = FALSE}, the default) or descending
order (\code{descending = TRUE}).}

\item{min_freq}{Specifies the minimum frequency that an ngram must have
in order to be included in the frequency list. With
\code{min_freq = 1} (the default), all ngrams are included.}

\item{word_char}{If \code{word_char = NULL} (the default), the user's locale
is used to distinguish word characters (e.g., "abc") from non-word
characters (e.g., ".?!"). A user's locale can be determined with
\code{sessionInfo()}. If words are split that shouldn't be, users can
give a character class to specify word characters. For example,
\code{word_char = "[-'a-z]+"} specifies that a combination of one or
more contiguous dashes, apostrophes or letters "a" to "z" be considered
as words, and as a result, the sentence "It's a hard-knock life,
for us!" has six words rather than more.}
}
\value{
A two-column local data frame, the first column with the
    ngrams and the second column with the frequencies.
}
\description{
Produces a frequency list of ngrams and returns a data frame. Users specify
the number of words in the ngrams, the order of the frequency list as either
alphabetical or by frequency, and whether the list be in ascending or
descending order, among other options.
}
\examples{
text <- c("First sentence here. Short, but sweet.")
text <- c(text, "Second one here, maybe?")
text <- c(text, "Third short paragraph here!")
text <- c(text, "Here too; with another thought.", "Here too.")

ngram_freq(text)
ngram_freq(text, ignore_case = FALSE)
ngram_freq(text, order_by = "freq", descending = TRUE)
ngram_freq(text, order_by = "freq", descending = TRUE, min_freq = 2)

# view difference (if any, given your locale)
ngram_freq("It's a hard-knock life, for us!")
ngram_freq("It's a hard-knock life, for us!", word_char = "[-'a-z]+")

# gets bigram frequencies
ngram_freq(text, num_wd = 2)
ngram_freq(text, num_wd = 2, order_by = "freq", descending = TRUE)

# gets trigram frequencies
ngram_freq(text, num_wd = 3)

}
\seealso{
For more info about local data frames, see
    \url{https://cran.r-project.org/web/packages/dplyr/vignettes/data_frames.html}).
}

