% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_freq.R
\name{word_freq}
\alias{word_freq}
\title{Produce a frequency list of words.}
\usage{
word_freq(text, ignore_case = TRUE, order_by = "alpha",
  descending = FALSE, min_freq = 1, word_char = NULL)
}
\arguments{
\item{text}{The text with the words whose frequencies are to be determined, as either a character vector, or something coercible to it.}

\item{ignore_case}{Specifies whether the frequency list is
case-sensitve or case-insensitive (the default). If case-insensitive
(\code{ignore_case = TRUE}) the words are converted to upper-case.}

\item{order_by}{Specifies whether the frequency list is ordered alphabetically
or by frequency, with the options \code{"alpha"} and \code{"freq"}.}

\item{descending}{Specifies whether the frequency list be ordered in
ascending order (the default) or descending order.}

\item{min_freq}{Specifies the minimum frequency that a word must have
in order to be included in the frequency list. With
\code{min_freq = 1} (the default), all words are included.}

\item{word_char}{If \code{word_char = NULL} (the default), the user's locale,
viewable with \code{sessionInfo()}, is used to distinguish word characters
(e.g., "abc") from non-word characters (e.g., ".?!"). If words are split
that shouldn't be, users can give a character class to specify word
characters. For example, \code{word_char = "[-'a-z]+"} specifies that a
combination of one or more contiguous dashes, apostrophes or letters
"a" to "z" be considered word characters, and as a result, the sentence
"It's a hard-knock life, for us!" has six words rather than more.}
}
\value{
Returns a two-column local data frame (see
    \url{https://cran.r-project.org/web/packages/dplyr/vignettes/data_frames.html},
    the first column with the words and the second column with the frequencies.
}
\description{
Produces a frequency list of words and returns a data frame. Users specify
the order of the frequency list as either alphabetical or by
frequency as well as whether the order by ascending or descending.
}
\examples{
text <- c("First sentence here. Short, but sweet.", "Second one here, maybe?",
"Third short paragraph here!", "Here too; with another thought.")
word_freq(text)
word_freq(text, ignore_case = FALSE)
word_freq(text, order_by = "freq", descending = TRUE)
word_freq(text, order_by = "freq", descending = TRUE, min_freq = 2)

# view difference (if any, given your locale) with \\code{word_char}
word_freq("It's a hard-knock like, for us!")
word_freq("It's a hard-knock like, for us!", word_char = "[-'a-z]+")
}

