% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord.R
\name{concord}
\alias{concord}
\title{Produce a concordance.}
\usage{
concord(text, pattern, ignore_case = TRUE, locations = FALSE,
  extra_context = 0, shorten = 0)
}
\arguments{
\item{text}{The text to search in, as either a character vector or a list
of character vectors.}

\item{pattern}{The Perl-compartible regular expression to search for.}

\item{ignore_case}{If \code{ignore_case = TRUE} (the default),
a case-insensitive search is performed.}

\item{locations}{If \code{locations = TRUE}, the element indexes with
matches are returned.}

\item{extra_context}{If positive, specifies the number of additional \strong{lines}
around the lines with matches to return. Useful if not enough context is
given in the lines with matches.}

\item{shorten}{If positive, truncates the preceding and following contexts
to the number of \strong{characters} specified. Useful if too much context is
given in the lines with matches, for example, in texts with (very)
long paragraphs.}
}
\value{
A data frame.
}
\description{
Produces a concordance of a regular expression and returns
a data frame.
}
\examples{
concord("hello", "e")
concord("some sentence here", "sentence")
concord("some sentence here", "s[eo]")

text <- c("first sentence here", "short paragraph here", "third line here")
concord(text, "e[nr]", locations = TRUE)
concord("some sentence here", "e(?=r)")
concord("some sentence here", "e(?!n)")
}
\seealso{
Stefan Th. Gries' function \code{\href{http://www.linguistics.ucsb.edu/faculty/stgries/exact_matches.r}{exact.matches()}}.
}

