% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpling.R
\docType{package}
\name{corpling}
\alias{corpling}
\alias{corpling-package}
\title{corpling: A package with some tools for corpus linguists}
\description{
The corpling package provides a simple concordancer and other tools.
}
\section{corpling functions}{

concord
}

